make float take precedence over int

+++ core/dbt/artifacts/resources/v1/components.py
@@ -17,7 +17,7 @@ from dbt_semantic_interfaces.type_enums import (
     TimeGranularity,
 )
 
-NodeVersion = Union[int, float, str]
+NodeVersion = Union[float, int, str]
 
 
 def _backcompat_doc_blocks(doc_blocks: Any) -> List[str]:
